https://github.com/git-commit-id/git-commit-id-maven-plugin/issues/562: reorgnaize project structure and move tests to either core or a better meaningfull location
