+++ src/test/java/pl/project13/maven/jgit/JgitProviderAheadBehindTest.java
@@ -15,19 +15,22 @@
  * along with git-commit-id-maven-plugin.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.project13.core;
+package pl.project13.maven.jgit;
 
+import pl.project13.core.JGitProvider;
 import pl.project13.maven.git.AheadBehindTest;
 
-public class NativeProviderAheadBehindTest extends AheadBehindTest<NativeGitProvider> {
+import java.nio.file.Paths;
+
+public class JgitProviderAheadBehindTest extends AheadBehindTest<JGitProvider> {
 
   @Override
-  protected NativeGitProvider gitProvider() {
-    return new NativeGitProvider(localRepository.getRoot(), 1000L, null);
+  public void extraSetup() {
+    gitProvider.setRepository(localRepositoryGit.getRepository());
   }
 
   @Override
-  protected void extraSetup() {
-    gitProvider.setEvaluateOnCommit("HEAD");
+  protected JGitProvider gitProvider() {
+    return JGitProvider.on(Paths.get(localRepository.getRoot().getAbsolutePath(), ".git").toFile(), null);
   }
 }
